Add display of transportation price
