+++ app/views/cargos/index.html.erb
@@ -16,9 +16,13 @@
           <td><%= link_to "#{cargo.id}", request_cargo_path(cargo.request_id, cargo) %></td>
           <td><%= cargo.weight %></td>
           <td><%= cargo.created_at %></td>
-          <td>NOT READY</td>
+          <td><%= @price[cargo.id] %></td>
         </tr>
       <% end %>
+      <td></td>
+      <td></td>
+      <td></td>
+      <td>total cost: <%= @price.values.inject { |a, b| a + b } %> </td>
     </table>
   </div>
 <% else %>
