+++ app/views/cargos/show.html.erb
@@ -1,5 +1,5 @@
 <br>
-<h3><%= link_to "<<< back to request", request_path(@request) %></h3>
+<h3><%= link_to "<<< back to list of cargos", request_cargos_path(@request) %></h3>
 <br>
 
 <p>Info about cargo where id = <%= @cargo.id %></p>
@@ -7,6 +7,5 @@
 <p>length : <%= @cargo.length %></p>
 <p>width  : <%= @cargo.width %></p>
 <p>height : <%= @cargo.height %></p>
-<p>Ratio per kilometer : <%= @coefficient_for_cargo %> монет</p>
 <br>
 <p>The price for the transportation of this particular cargo is <%= @price %> монет</p>
