+++ app/views/requests/show.html.erb
@@ -7,6 +7,8 @@
 <p>distance: <%= @request.distance %></p>
 <p>created at: <%= @request.created_at %></p>
 <br>
+<p>price: <%= @price %> монет</p>
+<br>
 <%= button_to 'Show cargos', request_cargos_path(@request), :method => :get  %>
 <br>
 <%= button_to 'Create cargo', new_request_cargo_path(@request), :method => :get  %>
